+++ lib/appstate_container.dart
@@ -273,29 +273,39 @@ class StateContainerState extends State<StateContainer> {
     });
   }
 
+  void updateCurrency(AvailableCurrency currency) {
+    setState(() {
+      curCurrency = currency;
+      currencyLocale = currency.getLocale().toString();
+    });
+    sl.get<MetadataService>().setCurrency(currency);
+  }
+
   Future<void> updateSolids() async {
-    if (wallet != null && wallet!.address != null && Address(wallet!.address).isValid()) {
-      final List<TXData> solids = await sl.get<DBHelper>().getAccountSpecificSolids(wallet!.address);
-      // check for duplicates and remove:
-      final Set<String?> uuids = <String?>{};
-      final List<int?> idsToRemove = <int?>[];
-      for (final TXData solid in solids) {
-        if (!uuids.contains(solid.uuid)) {
-          uuids.add(solid.uuid);
-        } else {
-          log.d("detected duplicate TXData! removing...");
-          idsToRemove.add(solid.id);
-          await sl.get<DBHelper>().deleteTXDataByID(solid.id);
-        }
-      }
-      for (final int? id in idsToRemove) {
-        solids.removeWhere((TXData element) => element.id == id);
-      }
-      setState(() {
-        wallet!.solids = solids;
-      });
-      EventTaxiImpl.singleton().fire(PaymentsHomeEvent(items: wallet!.solids));
+    if (wallet?.address == null || !Address(wallet!.address).isValid()) {
+      return;
     }
+
+    final List<TXData> solids = await sl.get<DBHelper>().getAccountSpecificSolids(wallet!.address);
+    // check for duplicates and remove:
+    final Set<String?> uuids = <String?>{};
+    final List<int?> idsToRemove = <int?>[];
+    for (final TXData solid in solids) {
+      if (!uuids.contains(solid.uuid)) {
+        uuids.add(solid.uuid);
+      } else {
+        log.d("detected duplicate TXData! removing...");
+        idsToRemove.add(solid.id);
+        await sl.get<DBHelper>().deleteTXDataByID(solid.id);
+      }
+    }
+    for (final int? id in idsToRemove) {
+      solids.removeWhere((TXData element) => element.id == id);
+    }
+    setState(() {
+      wallet!.solids = solids;
+    });
+    EventTaxiImpl.singleton().fire(PaymentsHomeEvent(items: wallet!.solids));
   }
 
   Future<void> updateTXMemos() async {
@@ -463,6 +473,7 @@ class StateContainerState extends State<StateContainer> {
       setState(() {
         currencyLocale = currency.getLocale().toString();
         curCurrency = currency;
+        sl.get<MetadataService>().setCurrency(currency);
       });
     });
     // Get default language setting
@@ -555,11 +566,9 @@ class StateContainerState extends State<StateContainer> {
     });
     _priceEventSub = EventTaxiImpl.singleton().registerTo<PriceEvent>().listen((PriceEvent event) {
       // PriceResponse's get pushed periodically, it wasn't a request we made so don't pop the queue
-      // handle the null case in debug mode:
       if (wallet != null) {
         setState(() {
           wallet!.localCurrencyPrice = event.response!.price?.toString() ?? wallet!.localCurrencyPrice;
-          wallet!.xmrPrice = event.response!.xmrPrice.toString();
         });
       }
     });
@@ -999,7 +1008,9 @@ class StateContainerState extends State<StateContainer> {
     EventTaxiImpl.singleton().fire(ConfirmationHeightChangedEvent(confirmationHeight: response.confirmationHeight));
 
     // check subscriptions:
-    sl.get<SubscriptionService>().checkAreSubscriptionsPaid(context);
+    if (wallet != null && wallet!.history != null && wallet!.history.isNotEmpty) {
+      sl.get<SubscriptionService>().checkAreSubscriptionsPaid(wallet!.history);
+    }
 
     setState(() {
       wallet!.loading = false;
@@ -1014,8 +1025,9 @@ class StateContainerState extends State<StateContainer> {
       } else {
         wallet!.accountBalance = BigInt.tryParse(response.balance!)!;
       }
-      wallet!.localCurrencyPrice = response.price.toString();
-      wallet!.xmrPrice = response.xmrPrice.toString();
+      if (response.price != null) {
+        wallet!.localCurrencyPrice = response.price.toString();
+      }
       sl.get<AccountService>().pop();
       sl.get<AccountService>().processQueue();
     });