+++ lib/ui/home_page.dart
@@ -7,6 +7,7 @@ import 'dart:math';
 
 import 'package:app_tracking_transparency/app_tracking_transparency.dart';
 import 'package:auto_size_text/auto_size_text.dart';
+import 'package:badges/badges.dart';
 import 'package:confetti/confetti.dart';
 import 'package:event_taxi/event_taxi.dart';
 import 'package:firebase_messaging/firebase_messaging.dart';
@@ -23,6 +24,7 @@ import 'package:wallet_flutter/bus/blocked_modified_event.dart';
 import 'package:wallet_flutter/bus/deep_link_event.dart';
 import 'package:wallet_flutter/bus/events.dart';
 import 'package:wallet_flutter/bus/payments_home_event.dart';
+import 'package:wallet_flutter/bus/subs_changed_event.dart';
 import 'package:wallet_flutter/bus/tx_update_event.dart';
 import 'package:wallet_flutter/bus/unified_home_event.dart';
 import 'package:wallet_flutter/bus/xmr_event.dart';
@@ -193,6 +195,7 @@ class AppHomePageState extends State<AppHomePage> with WidgetsBindingObserver, T
   int _selectedIndex = 2;
 
   bool _isPro = false;
+  List<Subscription> _subscriptions = [];
 
   Future<void> _switchToAccount(String account) async {
     final List<Account> accounts = await sl.get<DBHelper>().getAccounts(await StateContainer.of(context).getSeed());
@@ -898,6 +901,7 @@ class AppHomePageState extends State<AppHomePage> with WidgetsBindingObserver, T
 
       final ws = await sl.get<DBHelper>().getSelectedWorkSource();
       if (ws.type == WorkSourceTypes.URL) {
+        if (!mounted) return;
         StateContainer.of(context).stopLoading();
       }
 
@@ -906,7 +910,9 @@ class AppHomePageState extends State<AppHomePage> with WidgetsBindingObserver, T
       if (!mounted) return;
       Future<void>.delayed(const Duration(seconds: 5), () async {
         if (!mounted) return;
-        await sl.get<SubscriptionService>().checkAreSubscriptionsPaid(context);
+        if (StateContainer.of(context).wallet?.history.isNotEmpty ?? false) {
+          await sl.get<SubscriptionService>().checkAreSubscriptionsPaid(StateContainer.of(context).wallet!.history);
+        }
       });
     });
     // confetti:
@@ -1164,6 +1170,7 @@ class AppHomePageState extends State<AppHomePage> with WidgetsBindingObserver, T
   StreamSubscription<DeepLinkEvent>? _deepLinkEventSub;
   StreamSubscription<XMREvent>? _xmrSub;
   StreamSubscription<ConnStatusEvent>? _connectionSub;
+  StreamSubscription<SubsChangedEvent>? _subscriptionsSub;
   // purchase sub:
   StreamSubscription<List<PurchaseDetails>>? _subscription;
 
@@ -1305,6 +1312,11 @@ class AppHomePageState extends State<AppHomePage> with WidgetsBindingObserver, T
     if (Platform.isIOS) {
       InAppPurchase.instance.restorePurchases();
     }
+    _subscriptionsSub = EventTaxiImpl.singleton().registerTo<SubsChangedEvent>().listen((SubsChangedEvent event) {
+      setState(() {
+        _subscriptions = event.subs ?? [];
+      });
+    });
   }
 
   void _destroyBus() {
@@ -2547,6 +2559,12 @@ class AppHomePageState extends State<AppHomePage> with WidgetsBindingObserver, T
   }
 
   Widget _buildBottomNavigationBar(BuildContext context) {
+    int unpaidSubCount = 0;
+    for (final Subscription sub in _subscriptions) {
+      if (sub.active && !sub.paid) {
+        unpaidSubCount++;
+      }
+    }
     return Container(
       decoration: const BoxDecoration(
         borderRadius: BorderRadius.only(topRight: Radius.circular(30), topLeft: Radius.circular(30)),
@@ -2568,7 +2586,13 @@ class AppHomePageState extends State<AppHomePage> with WidgetsBindingObserver, T
               backgroundColor: StateContainer.of(context).curTheme.warning,
             ),
             BottomNavigationBarItem(
-              icon: const Icon(Icons.currency_exchange),
+              icon: Badge(
+                showBadge: unpaidSubCount > 0,
+                badgeContent: Text("$unpaidSubCount", style: const TextStyle(color: Colors.white)),
+                animationType: BadgeAnimationType.scale,
+                shape: BadgeShape.circle,
+                child: const Icon(Icons.currency_exchange),
+              ),
               label: Z.of(context).subsButton,
               backgroundColor: StateContainer.of(context).curTheme.warning,
             ),